feat(nf1,firmware): misc updates to firmware for events (previous floor, alarms, calibration on live,  avoid recalibrating with each env var change.)
sends an event about the current floor at 5 minute intervals when idle, and 2 minute intervals when live.
events include previousFloor previousFloorChanged which denotes what the previous floor was and when it was changed.
events include direction which is 0 for no change, or a positive or negative value indicating the number of floors changed.
when a device becomes live, it calibrates the baseline floor and pressure.
The floor calculation is offset to avoid boundary conditions. A the floor extends 30% of the floor height below the actual floor level and 70% above it.
a no-movement alarm is sent when the device is live and no floor change has happened within the no movement threshold (which is now given in seconds, rather than minutes.) The alarm is sent just once. The data.qo event has an alarm property. The alarm is reset if live is toggled, or the baseline floor changed.