Mdm/firmware alarms (#8)
* feat(nf1,firmware): adds debugging support for stlink

* feat(nf1,firmware): misc updates to firmware for events (previous floor, alarms, calibration on live,  avoid recalibrating with each env var change.)

sends an event about the current floor at 5 minute intervals when idle, and 2 minute intervals when live.
events include previousFloor previousFloorChanged which denotes what the previous floor was and when it was changed.
events include direction which is 0 for no change, or a positive or negative value indicating the number of floors changed.
when a device becomes live, it calibrates the baseline floor and pressure.
The floor calculation is offset to avoid boundary conditions. A the floor extends 30% of the floor height below the actual floor level and 70% above it.
a no-movement alarm is sent when the device is live and no floor change has happened within the no movement threshold (which is now given in seconds, rather than minutes.) The alarm is sent just once. The data.qo event has an alarm property. The alarm is reset if live is toggled, or the baseline floor changed.

* feat(nf1,firmware): bump minor version

* chore(nf1,firmware): factor out reading and app state. The intent was to move more of the state processing code out of the main app so that it can be unit tested.

* feat: adds a FIR moving average filter to smooth out floor changes.  Publish a floor change explicitly when the baseline is set.

* small tweaks: remove `prevFloorChange` in events (it's device local time, not epoch time - was just intended for debugging.)  Fixed a typo,  adds `app`:`nf1` to the event payload.

* uses `floor.qo` and `alarm.qo` as given in the spec.

* use "c/c++" formatter on main.cpp

* format with c/cpp formatting

Co-authored-by: Carlton Henderson <henderson.carlton@gmail.com>
Co-authored-by: TJ VanToll <tj.vantoll@gmail.com>