Zak NF09 (#97)
* feat: Zephyr devcontainer

* chore: use documented API variant

* fix: Flow rate calculation

* chore: clear dead whitespace

* chore: Improve boundary flow rate calc

* nf9: Make various improvements to the Zephyr implementation.

- Use // for comments (more IDE friendly).
- Add Notecard lock/unlock callbacks so that various asynchronous operations
(e.g. note.add and environment variable updating) don't collide with each other
and screw up bus traffic. Also, this is a nice demo of how to use note-c's
locking functionality.
- Call attnArm after processing the data.qi notes rather than before.
- Instead of checking the alarm conditions every iteration of the main loop, do
it every 10 seconds.
- To prevent a bunch of alarms from rattling off one after the other (for
example, if the flow rate is popping above the upper limit and falling below it
rapidly), add an alarm "cooldown timer" so that we publish an alarm a max of
once every 30 seconds.
- Make a bunch of AppState variables volatile, as they're now modified in timer
callbacks, which are called from ISRs.

* chore: address code review feedback

* chore: throttle msgs to change

* fix: reset timers on valve toggle

* chore: refactor notecard to note_c_hooks

* fix: refactor valve publish requirement

---------

Co-authored-by: Hayden Roche <hroche@blues.com>