Retry hub.set in various apps' firmware.
This addresses a problem where the Notecard and host have just booted up and the
Notecard isn't ready to receive the first request.

This change touched several different apps, so I added a script to build all
these apps' firmware, which happen to all be PlatformIO-based. In doing this,
I found some other problems:

- NF10 firmware was broken because the APP_NAME macro wasn't quoted. Also, the
symlink to NF9's main.cpp was broken.
- NF10 and NF35 needed a newer version of note-arduino so they could use
sendRequestWithRetry.